requirements.txt
aiogram
fastapi
uvicorn
jinja2
python-dotenv
aiosqlite
python-multipart
requests
pystray
Pillow
